Refactor map saver.
