+++ lmeditor/src/app/app.cpp
@@ -1,4 +1,5 @@
 #include "app.h"
+#include "saver.h"
 #include <boost/rational.hpp>
 #include <lmeditor/map_editor.h>
 #include <lmlib/variant_visitor.h>
@@ -290,5 +291,43 @@ void editor_app::init_command_help()
             dynamic_cast<command_help *>(help_widget)->add_to_frame(frame);
         },
     };
+}
+
+void editor_app::init_map_saver()
+{
+    auto path = map_file_project_relative_path.string();
+    state.emplace<modal_state>(modal_state{
+      .modal = std::make_unique<saver>(
+        resources, "Save map", path.substr(0, path.rfind(".lmap"))),
+      .input_handler =
+        [](editor_app &app, auto widget, auto &input_event) {
+            auto saver = dynamic_cast<lmeditor::saver *>(widget);
+            auto maybe_key_down_msg =
+              std::get_if<lmtk::key_down_event>(&input_event);
+            if (
+              maybe_key_down_msg &&
+              maybe_key_down_msg->key == lmpl::key_code::Enter)
+            {
+                auto relative = saver->field.get_value();
+                auto absolute = app.project_dir / (relative + ".lmap");
+
+                app.save_map(absolute);
+
+                app.map_file_project_relative_path = relative;
+                app.state.emplace<gui_state>(app);
+                return true;
+            }
+
+            return saver->field.handle(
+              input_event,
+              app.resources.renderer.get(),
+              app.resources.font.get(),
+              app.resources.resource_sink);
+        },
+      .renderer =
+        [](auto saver, auto frame) {
+            dynamic_cast<lmeditor::saver *>(saver)->add_to_frame(frame);
+        },
+    });
 }
 } // namespace lmeditor
