+++ lmeditor/src/app/states/gui.cpp
@@ -1,5 +1,5 @@
 #include "../app.h"
-#include "../map_saver.h"
+#include "../saver.h"
 #include <lmeditor/map_selector.h>
 #include <lmtk/char_field.h>
